again

+++ go-evdev/device.go
@@ -286,8 +286,8 @@ func (d *InputDevice) UploadEffect(effect Effect) error {
 	err = d.WriteOne(&InputEvent{
 		Time:  timeVal,
 		Type:  EV_FF,
-		Code:  val,
-		Value: val,
+		Code:  uint16(val),
+		Value: int32(val),
 	})
 	slog.Info("c output", "return", val, "error", err)
 	return err
